rag-pipeline/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
openpyxl==3.1.2

# Text processing
langdetect==1.0.9

# Embeddings and vector store
sentence-transformers==2.6.0
torch==2.2.0
numpy==1.26.4

# Vector stores (choose one or both)
faiss-cpu==1.8.0
qdrant-client==1.7.0  

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1